Test Download v1

+++ app/src/main/java/com/d0st/bhadoozindex/dto/Cdn.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Cdn(
     val name:String,
-    val size:Int,
+    val size:Long,
     val parts:Int,
     val part_details:Map<String, String>,
     val sha256:String
